HLint: ignore all warnings by default
We previously disabled HLint in Travis because we thought there was no good
way to use it without spending time growing the list of disabled hints. I
recently learned that HLint has a whitelist mode, which I've turned on here.

I've enabled a small set of warnings that we can hopefully all agree on, and
turned HLint back on in Travis by default. There are a few TODOs from HLint
bugs remaining.

The TODOs on our side are:
- Fix the legitimate warnings that are commented out in this PR
- Brainstorm custom warnings for better code quality