uc-crux-llvm: Don't give up when there are new preconditions
Previously, the bugfinding loop would give up whenever it encountered an
error it didn't know how to classify. This patch makes it so that it
only gives up if that's true *and* it couldn't deduce new preconditions
from any of the errors it encountered. This is a good idea because there
might be multiple errors that share the same root cause, which is
resolved by the new preconditions.

The included program is an example of this phenomenon.