uc-crux-llvm: Check programs for "crash-equivalence"
This commit adds another feature from the original paper
"Under-Constrained Symbolic Execution: Correctness Checking for Real
Code" by Ramos and Engler, namely the ability to check two versions of
the same function for "crash equivalence" or "crash ordering", where one
version is "crash less than" the other just if the all the undefined
behaviors present in the former are also present in the latter.

The idea is that this can be used e.g. to test patches: patches should
generally only *remove* instances of undefined behavior, not add them.

This can also be used as a generic, automated "rough equivalence" check.